Repo structure restored
Fixed problems with repo structure
Additional method for detection of SmartIPs
