+++ custom_components/habitron/manifest.json
