+++ custom_components/habitron/strings.json
@@ -1,0 +1,63 @@
+{
+  "config": {
+    "abort": {
+      "already_configured": "Already_configured_device"
+    },
+    "error": {
+      "cannot_connect": "Cannot connect to SmartIP",
+      "host_not_found": "Host name not fount, try with IP",
+      "interval_too_long": "Interval too short, enter max. 10 s",
+      "interval_too_short": "Interval too short, enter min. 2 s",
+      "invalid_interval": "Invalid interval entry",
+      "unknown": "Unknown"
+    },
+    "step": {
+      "user": {
+        "data": {
+          "habitron_host": "Host name or IP of SmartIP",
+          "update_interval": "Update interval [s]"
+        },
+        "title": "Setup communication"
+      }
+    }
+  },
+  "options": {
+    "abort": {
+      "already_configured": "Already_configured_device"
+    },
+    "error": {
+      "cannot_connect": "Cannot connect to SmartIP",
+      "host_not_found": "Host name not fount, try with IP",
+      "interval_too_long": "Interval too short, enter max. 10 s",
+      "interval_too_short": "Interval too short, enter min. 2 s",
+      "invalid_interval": "Invalid interval entry",
+      "unknown": "Unknown"
+    },
+    "step": {
+      "init": {
+        "data": {
+          "habitron_host": "Host name or IP of SmartIP",
+          "update_interval": "Update interval [s]",
+          "updates_enabled": "Perform cyclic updates"
+        },
+        "title": "Setup communication"
+      }
+    }
+  },
+  "entity": {
+    "select": {
+      "habitron_mode": {
+        "state": {
+          "On": "On",
+          "Off": "Off",
+          "Day": "Day",
+          "Night": "Night",
+          "Absent": "Absent",
+          "Present": "Present",
+          "Sleeping": "Sleeping",
+          "Vacation": "Vacation"
+        }
+      }
+    }
+  }
+}
