+++ hacs.json
@@ -1,0 +1,5 @@
+{
+  "name": "Habitron",
+  "iot_class": "Local Pull",
+  "homeassistant": "2023.1.1"
+}
