--- services.yaml
@@ -1,127 +1,0 @@
-# Describes the format for Habitron service
-mod_restart:
-  name: Module restart
-  description: Restart one or all Habitron modules
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
-    mod_nmbr:
-      name: Module number
-      description: Habitron's module address, default 0xFF for all modules
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 64
-rtr_restart:
-  name: Router restart
-  description: Restart the Habitron router
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
-save_module_smc:
-  name: Save module smc file
-  description: Get module definitions of rules and names and save them to .smc text file
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
-    mod_nmbr:
-      name: Module number
-      description: Habitron's module address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 64
-save_module_smg:
-  name: Save module smg file
-  description: Get module settings and save them to .smg text file
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
-    mod_nmbr:
-      name: Module number
-      description: Habitron's module address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 64
-save_router_smr:
-  name: Save router smr file
-  description: Get router settings and save them to .smr text file
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
-save_module_status:
-  name: Save module status to file
-  description: Get module status and save it to .mstat text file
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
-    mod_nmbr:
-      name: Module number
-      description: Habitron's module address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 64
-save_router_status:
-  name: Save router status to file
-  description: Get router status and save it to .rstat text file
-  fields:
-    rtr_nmbr:
-      name: Router number
-      description: Habitron's router address
-      required: true
-      example: 1
-      selector:
-        number:
-          min: 1
-          max: 8
