--- translations/de.json
@@ -1,63 +1,0 @@
-{
-    "config": {
-        "step": {
-            "user": {
-                "title": "Einstellungen zur Kommunikation",
-                "data": {
-                    "habitron_host": "Host-Name oder IP des SmartIP",
-                    "update_interval": "Update Intervall [s]"
-                }
-            }
-        },
-        "error": {
-            "cannot_connect": "Keine Verbindung zum SmartIP",
-            "host_not_found": "Hostname nicht gefunden, evtl. IP eingeben",
-            "invalid_interval": "Intervall Eingabe ungültig",
-            "interval_too_short": "Intervall zu kurz, mind. 2 s",
-            "interval_too_long": "Intervall zu lang, max. 10 s",
-            "unknown": "Unbekannter Fehler"
-        },
-        "abort": {
-            "already_configured": "Habitron Integration bereits installiert"
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "title": "Einstellungen zur Kommunikation",
-                "data": {
-                    "habitron_host": "Host-Name oder IP des SmartIP",
-                    "update_interval": "Update Intervall [s]",
-                    "updates_enabled": "Zyklische Updates ausführen"
-                }
-            }
-        },
-        "error": {
-            "cannot_connect": "Keine Verbindung zum SmartIP",
-            "host_not_found": "Hostname nicht gefunden, evtl. IP eingeben",
-            "invalid_interval": "Intervall Eingabe ungültig",
-            "interval_too_short": "Intervall zu kurz, mind. 2 s",
-            "interval_too_long": "Intervall zu lang, max. 10 s",
-            "unknown": "Unbekannter Fehler"
-        },
-        "abort": {
-            "already_configured": "Habitron Integration bereits installiert"
-        }
-    },
-    "entity": {
-      "select": {
-        "habitron_mode": {
-            "state": {
-                "On": "Aktiv",
-                "Off": "Aus",
-                "Day": "Tag",
-                "Night": "Nacht",
-                "Absent": "Abwesend",
-                "Present": "Anwesend",
-                "Sleeping": "Schlafen",
-                "Vacation": "Urlaub"
-            }
-        }
-      }
-    }
-}
